java/src/test/resources/pii/requirements.txt
pip
setuptools
wheel
spacy==3.4.4
presidio-analyzer==2.2.*
presidio-anonymizer==2.2.*
lingua-language-detector
faker
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.4.1/en_core_web_lg-3.4.1-py3-none-any.whl